Rename _utils.py to _names.py
"utils" is non-informative — the module's actual concern is reserved
coordinate/variable names and name-or-index resolution. Naming it _names.py
says what's inside and what isn't. ADR-012 updated to reflect the rename and
justify the choice; CLAUDE.md example updated.
